month-01-python-sql-foundations/article-007-advanced-sql/requirements.txt
pandas==2.1.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
colorama==0.4.6
tabulate==0.9.0
matplotlib==3.8.2
seaborn==0.13.0